activity item to tsx
