+++ app/webpack/observations/show/components/activity_item.test.tsx
@@ -1,17 +1,18 @@
 import React from "react";
 import { render, screen } from "@testing-library/react";
+import type { ActivityItemModel, ActivityItemProps } from "./activity_item";
 
 // Stub shared components that transitively load browser-only deps (heic-to, etc.)
 jest.mock( "../../../taxa/shared/util", ( ) => ( {
-  urlForTaxon: t => ( t ? `/taxa/${t.id}` : null )
+  urlForTaxon: ( t: { id: number } | null ) => ( t ? `/taxa/${t.id}` : null )
 } ) );
 jest.mock( "../../../shared/components/split_taxon", ( ) => ( {
   __esModule: true,
-  default: ( { taxon } ) => <span>{ taxon?.name }</span>
+  default: ( { taxon }: { taxon?: { name?: string } } ) => <span>{ taxon?.name }</span>
 } ) );
 jest.mock( "../../../shared/components/user_text", ( ) => ( {
   __esModule: true,
-  default: ( { text } ) => <span>{ text }</span>
+  default: ( { text }: { text?: string } ) => <span>{ text }</span>
 } ) );
 jest.mock( "../../../shared/components/user_image", ( ) => ( {
   __esModule: true,
@@ -19,7 +20,7 @@ jest.mock( "../../../shared/components/user_image", ( ) => ( {
 } ) );
 jest.mock( "../../../shared/components/user_link", ( ) => ( {
   __esModule: true,
-  default: ( { user } ) => <span>{ user?.login }</span>
+  default: ( { user }: { user?: { login?: string } } ) => <span>{ user?.login }</span>
 } ) );
 jest.mock( "../../../shared/components/inativersary", ( ) => ( {
   __esModule: true,
@@ -48,7 +49,10 @@ jest.mock( "./users_popover", ( ) => ( {
 
 // jQuery: used in componentDidUpdate for textarea mentions and inline in render
 // for OverlayTrigger container ($(...).get(0)).
-global.$ = jest.fn( ).mockReturnValue( { textcompleteUsers: jest.fn( ), get: ( ) => null } );
+( global as unknown as { $: unknown } ).$ = jest.fn( ).mockReturnValue( {
+  textcompleteUsers: jest.fn( ),
+  get: ( ) => null
+} );
 
 // eslint-disable-next-line import/first
 import ActivityItem from "./activity_item";
@@ -74,6 +78,10 @@ const baseItem = {
   firstDisplay: false
 };
 
+const itemWith = ( overrides: Record<string, unknown> ): ActivityItemModel => (
+  { ...baseItem, ...overrides } as unknown as ActivityItemModel
+);
+
 const baseProps = {
   observation: { id: 99, user: { id: 20, login: "bob" }, identifications: [] },
   config: {
@@ -81,9 +89,15 @@ const baseProps = {
     currentUserCanInteractWithResource: ( ) => false
   },
   item: baseItem,
+  addID: noop,
+  editID: noop,
+  editComment: noop,
+  confirmDeleteID: noop,
   performOrOpenConfirmationModal: noop,
-  setFlaggingModalState: noop
-};
+  setFlaggingModalState: noop,
+  voteIdentification: noop,
+  unvoteIdentification: noop
+} as unknown as ActivityItemProps;
 
 describe( "ActivityItem", ( ) => {
   it( "renders an identification item with its taxon name", ( ) => {
@@ -92,13 +106,13 @@ describe( "ActivityItem", ( ) => {
   } );
 
   it( "renders a comment item with its body text", ( ) => {
-    const item = { ...baseItem, taxon: null, body: "Nice find!" };
+    const item = itemWith( { taxon: null, body: "Nice find!" } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByText( "Nice find!" ) ).toBeInTheDocument( );
   } );
 
   it( "maverick status badge has title attribute", ( ) => {
-    const item = { ...baseItem, category: "maverick" };
+    const item = itemWith( { category: "maverick" } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTitle( "maverick" ) ).toBeInTheDocument( );
     // Label text is wrapped in item-status-label span for responsive hiding.
@@ -106,31 +120,31 @@ describe( "ActivityItem", ( ) => {
   } );
 
   it( "improving status badge has title attribute", ( ) => {
-    const item = { ...baseItem, category: "improving" };
+    const item = itemWith( { category: "improving" } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTitle( "improving" ) ).toBeInTheDocument( );
   } );
 
   it( "leading status badge has title attribute", ( ) => {
-    const item = { ...baseItem, category: "leading" };
+    const item = itemWith( { category: "leading" } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTitle( "leading" ) ).toBeInTheDocument( );
   } );
 
   it( "withdrawn status badge has title attribute when identification is not current", ( ) => {
-    const item = { ...baseItem, current: false };
+    const item = itemWith( { current: false } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTitle( "id_withdrawn" ) ).toBeInTheDocument( );
   } );
 
   it( "flagged status badge has title attribute when item has unresolved flags", ( ) => {
-    const item = { ...baseItem, flags: [{ resolved: false }] };
+    const item = itemWith( { flags: [{ resolved: false }] } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTitle( "flagged_" ) ).toBeInTheDocument( );
   } );
 
   it( "renders HiddenActivityItem when item is hidden and viewer cannot see hidden content", ( ) => {
-    const item = { ...baseItem, hidden: true };
+    const item = itemWith( { hidden: true } );
     render( <ActivityItem {...baseProps} item={item} /> );
     expect( screen.getByTestId( "hidden-activity-item" ) ).toBeInTheDocument( );
   } );
